libwaku-cpp-example: move base64 funcs to a separate file
