+++ examples/cpp/waku.cpp
@@ -13,6 +13,7 @@
 #include <unistd.h>
 #include <sys/syscall.h>
 
+#include "base64.h"
 #include "../../library/libwaku.h"
 
 #define WAKU_CALL(call)                                                        \
@@ -78,58 +79,6 @@ static error_t parse_opt(int key, char *arg, struct argp_state *state) {
 }
 
 static struct argp argp = { options, parse_opt, args_doc, doc, 0, 0, 0 };
-
-// Base64 encoding
-// source: https://nachtimwald.com/2017/11/18/base64-encode-and-decode-in-c/
-size_t b64_encoded_size(size_t inlen)
-{
-	size_t ret;
-
-	ret = inlen;
-	if (inlen % 3 != 0)
-		ret += 3 - (inlen % 3);
-	ret /= 3;
-	ret *= 4;
-
-	return ret;
-}
-
-const char b64chars[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
-
-void b64_encode(char* in, size_t len, std::vector<char>& out)
-{
-	size_t  elen;
-	size_t  i;
-	size_t  j;
-	size_t  v;
-
-	if (in == NULL || len == 0)
-		return;
-
-	elen = b64_encoded_size(len);
-	out.reserve(elen+1);
-
-	for (i=0, j=0; i<len; i+=3, j+=4) {
-		v = in[i];
-		v = i+1 < len ? v << 8 | in[i+1] : v << 8;
-		v = i+2 < len ? v << 8 | in[i+2] : v << 8;
-
-		out[j]   = b64chars[(v >> 18) & 0x3F];
-		out[j+1] = b64chars[(v >> 12) & 0x3F];
-		if (i+1 < len) {
-			out[j+2] = b64chars[(v >> 6) & 0x3F];
-		} else {
-			out[j+2] = '=';
-		}
-		if (i+2 < len) {
-			out[j+3] = b64chars[v & 0x3F];
-		} else {
-			out[j+3] = '=';
-		}
-	}
-}
-
-// End of Base64 encoding
 
 // Beginning of UI program logic
